requirements.txt
anthropic>=0.50.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
openpyxl>=3.1.0
pydantic>=2.0
python-dotenv>=1.0.0
PyYAML>=6.0
rich>=13.0
